requirements.txt
aiohttp==3.12.13
beautifulsoup4==4.12.1
chromadb==1.3.0
ddgs==9.5.5
docx2txt==0.9
duckduckgo-search==7.5.0
fastmcp==2.10.5
html2text==2025.4.15
langchain==1.0.3
langchain-chroma==1.0.0
langchain-community==0.4.1
langchain-mcp-adapters==0.1.12
langchain-openai==1.0.1
langgraph==1.0.2
langgraph-supervisor==0.0.30
lark==1.2.2
lxml==6.0.0
notebook==7.4.5
openai==2.6.1
pypdf==6.0.0
python-dotenv==1.1.1
requests==2.32.5
tiktoken==0.11.0
wikipedia==1.4.0
